make attributes semantically more rich
positional attributes are similar to before, they are parsed but now the
raw string is also available

named attirbutes are now either plain, wth unparsed strings,
or nested with a new attributes block as the value

nested attributes could be simulated before but in a roundabout way.
this is now more straightforward and should simplify the use case of
scim as a configuration language. this also makes it much more
straightforward to have keyed attributes with multiple values